requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
pillow>=10.2.0
zxing-cpp>=2.2.0
